fix(host-details): Change from accordion to tab
