--- tavern/internal/www/src/features/task-output/index.tsx
@@ -1,1 +1,0 @@
-export {TaskOutput} from "./TaskOutput";
